refactor(CardItem): separate CardItem file
